+++ src/extension/endpoint/index.ts
@@ -1,2 +1,3 @@
 export { HttpEndpoint } from './http-endpoint/http-endpoint';
+export { FileEndpoint } from './file-endpoint/file-endpoint';
 export { Endpoint } from './endpoint';