+++ src/extension/file-io/index.ts
@@ -1,1 +1,0 @@
-export { SparqlNotebookSerializer } from './sparql-notebook-serializer.class';