+++ src/extension/file-io/raw-notebook-cell.model.ts
@@ -1,8 +1,0 @@
-import * as vscode from "vscode";
-
-export interface RawNotebookCell {
-  language: string;
-  value: string;
-  kind: vscode.NotebookCellKind;
-  metadata?: any;
-}