+++ src/extension/notebook/file-io/index.ts
@@ -1,0 +1,1 @@
+export { SparqlNotebookSerializer } from './sparql-notebook-serializer.class';