+++ src/extension/notebook/file-io/raw-notebook-cell.model.ts
@@ -1,0 +1,8 @@
+import * as vscode from "vscode";
+
+export interface RawNotebookCell {
+  language: string;
+  value: string;
+  kind: vscode.NotebookCellKind;
+  metadata?: any;
+}